+++ nekro_agent/__init__.py
@@ -7,17 +7,14 @@ from nonebot.plugin import PluginMetadata
 from pydantic import BaseModel
 
 import nekro_agent.core.bot
-import nekro_agent.matchers
+from nekro_agent.adapters import load_adapters
 from nekro_agent.core.args import Args
 from nekro_agent.core.config import config
 from nekro_agent.core.database import init_db
 from nekro_agent.core.logger import logger
-from nekro_agent.models.db_chat_channel import DBChatChannel
 from nekro_agent.routers import mount_routers
 from nekro_agent.services.festival_service import festival_service
 from nekro_agent.services.mail.mail_service import send_bot_status_email
-
-# from nekro_agent.services.extension import init_extensions, reload_ext_workdir
 from nekro_agent.services.plugin.collector import init_plugins
 from nekro_agent.services.timer_service import timer_service
 from nekro_agent.systems.cloud.scheduler import start_telemetry_task
@@ -40,13 +37,12 @@ if config.WEAVE_ENABLED:
         logger.error(f"Weave 服务连接失败: {e}")
 
 mount_routers(get_app())
-# init_extensions()
-# reload_ext_workdir()
 
 
 @get_driver().on_startup
 async def on_startup():
     await init_db()
+    await load_adapters(get_app())
     await init_plugins()
     await timer_service.start()
     logger.info("Timer service initialized")
