+++ nekro_agent/adapters/nonebot/matchers/__init__.py
@@ -1,0 +1,5 @@
+from nekro_agent.core.logger import logger
+
+from . import command, message, notice
+
+logger.success("Matchers loaded successfully")
