--- nekro_agent/matchers/__init__.py
@@ -1,7 +1,0 @@
-import nekro_agent.matchers.command
-import nekro_agent.matchers.notice
-from nekro_agent.core.logger import logger
-
-from .message import message_matcher, upload_notice_matcher
-
-logger.success("Matchers loaded successfully")
