+++ nekro_agent/services/message/message_service.py
@@ -2,7 +2,6 @@ import asyncio
 import json
 import re
 import time
-from datetime import datetime
 from pathlib import Path
 from typing import Dict, List, Optional, Union
