+++ plugins/builtin/ai_voice.py
@@ -3,10 +3,14 @@ from nonebot.matcher import Matcher
 from nonebot.params import CommandArg
 from pydantic import Field
 
+from nekro_agent.adapters.nonebot.matchers.command import (
+    command_guard,
+    finish_with,
+    on_command,
+)
 from nekro_agent.api import context, core
 from nekro_agent.api.plugin import ConfigBase, NekroPlugin, SandboxMethodType
 from nekro_agent.api.schemas import AgentCtx
-from nekro_agent.matchers.command import command_guard, finish_with, on_command
 from nekro_agent.schemas.chat_message import ChatType
 
 plugin = NekroPlugin(
