+++ plugins/builtin/dice.py
@@ -3,12 +3,15 @@ import random
 from nonebot.adapters.onebot.v11 import Bot, Message, MessageEvent
 from nonebot.matcher import Matcher
 from nonebot.params import CommandArg
-from pydantic import Field
 
+from nekro_agent.adapters.nonebot.matchers.command import (
+    command_guard,
+    finish_with,
+    on_command,
+)
 from nekro_agent.api import core, message
 from nekro_agent.api.plugin import ConfigBase, NekroPlugin, SandboxMethodType
 from nekro_agent.api.schemas import AgentCtx
-from nekro_agent.matchers.command import command_guard, finish_with, on_command
 
 plugin = NekroPlugin(
     name="掷骰姬",
